Created new smbiz project in the existing tll/smbiz dir. Moved smbiz-server/confg and smbiz-server/build.gant up one level (the smbiz dir).
Converted ConfigKeys to an enum so we may iterate the keys!

Created new IConfigKeyProvider in the tll/config project to serve as a way to filter out unwanted config properties.  Namely to be used for creating the config.properties for the built smbiz-server
TODO: figure out how to dynamically load smbiz-server's ConfigKeys dynamically since this is the relevant IConfigKeyProvider for this task!