Re-factored IEntityFactory making it compatible with the likes of gaej.
Updated GaejPrimaryKeyGenerator to use the gaej dao impl.  Now the strategy for providing generated primary keys is to persist the entity and grab the resultant long id.  This is a bit inefficient but it should keep me from having to significantly re-factor other portions of my codebase.

Added some minor null value guarding in Marshaler.